Test packages now follows new convention
	Reviewed by: Simon

+++ test/domain/BookingTest.java
@@ -116,7 +116,7 @@ public class BookingTest {
 	System.out.println("getRent");
 	Booking instance = null;
 	int expResult = 0;
-	int result = instance.getRent();
+	double result = instance.getRent();
 	assertEquals(expResult, result);
 	// TODO review the generated test code and remove the default call to fail.
 	fail("The test case is a prototype.");
